chore: hotfix dep breaking changes

+++ FF16Tools.Pack/Packing/FF16PackBuilder.cs
@@ -390,7 +390,7 @@ public class FF16PackBuilder
             task.PackFile.DataOffset = (ulong)packStream.Position;
 
             using MemoryOwner<byte> decompBuffer = MemoryOwner<byte>.Allocate((int)task.PackFile.DecompressedFileSize);
-            long sizeCompressed = _codec.CompressBufferBound((long)task.PackFile.DecompressedFileSize * 2); // Incase
+            long sizeCompressed = _codec.CompressBufferBound(task.PackFile.DecompressedFileSize * 2); // Incase
             using MemoryOwner<byte> compBuffer = MemoryOwner<byte>.Allocate((int)sizeCompressed);
 
             using var fileStream = new FileStream(task.LocalPath, FileMode.Open);
